MDTZ-978 Addressed pr comments
