+++ src/infrastructure/figma/figma-service.ts
@@ -21,8 +21,6 @@ import type {
 	ConnectUserInfo,
 	FigmaDesignIdentifier,
 } from '../../domain/entities';
-import { parseJsonOfSchemaFromObject } from '../ajv';
-import { FILE_RESPONSE_SCHEMA } from '../jira/schemas';
 import { getLogger } from '../logger';
 
 const buildDevResourceNameFromJiraIssue = (
@@ -64,8 +62,6 @@ export class FigmaService {
 				},
 				accessToken,
 			);
-
-			parseJsonOfSchemaFromObject(fileResponse, FILE_RESPONSE_SCHEMA);
 
 			return transformNodeToAtlassianDesign({
 				fileKey: designId.fileKey,
@@ -116,8 +112,6 @@ export class FigmaService {
 		);
 
 		return designIds.map((designId) => {
-			parseJsonOfSchemaFromObject(fileResponse, FILE_RESPONSE_SCHEMA);
-
 			if (!designId.nodeId) {
 				return transformFileToAtlassianDesign({
 					fileKey: designId.fileKey,
